redesign application router: path, pages & layouts
