+++ src/app/[locale]/(authenticated)/_layout.tsx
