+++ src/app/[locale]/(authenticated)/_page.tsx
