+++ src/app/[locale]/(authenticated)/blog/_page.tsx
