+++ src/app/[locale]/(authenticated)/products/_page.tsx
