+++ src/app/[locale]/(authenticated)/profile/_page.tsx
