+++ src/app/[locale]/(landing)/about/page.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function AboutPage() {
+  return <div>About page</div>;
+}
