+++ src/app/[locale]/(landing)/contact/loading.tsx
