+++ src/app/[locale]/(landing)/page.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function HomePage() {
+  return <div>Home page</div>;
+}
