--- src/app/[locale]/_page.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Home() {
-  return <div>Landing static page</div>;
-}
