+++ src/app/[locale]/dashboard/register/loading.tsx
