--- src/app/[locale]/not-found.tsx
@@ -1,3 +1,0 @@
-export default function NotFound() {
-  return <div>Page not found</div>;
-}
